More robust PostgreSQL version detection
The version check only worked with pg_config reporting 3-component
version number *and* no extra information. This does not work with libpq
as available in Ubuntu repositories, which both (a) only reports major
and minor version components, and (b) adds more information in the
parens afterwards.

But only major and minor are actually used, so just read those in and
slurp the rest of the line.

Fixes #4.